added clippy, refacotred to more sensical file org
