+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+
 serde = {version = "1.0", features = ["derive"]}
 serde_json = "1.0.82"
 anyhow = "1.0"
