+++ clippy.toml
@@ -1,0 +1,1 @@
+#![deny(clippy::pedantic)]
